+++ src/main/java/insane96mcp/progressivebosses/mixin/EnderDragonEntityMixin.java
@@ -1,10 +1,13 @@
 package insane96mcp.progressivebosses.mixin;
 
+import insane96mcp.progressivebosses.module.dragon.phase.CrystalRespawnPhase;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.damagesource.EntityDamageSource;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.boss.EnderDragonPart;
 import net.minecraft.world.entity.boss.enderdragon.EnderDragon;
+import net.minecraft.world.entity.boss.enderdragon.phases.EnderDragonPhase;
 import net.minecraft.world.level.Level;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -22,6 +25,15 @@ public class EnderDragonEntityMixin extends Mob {
 		EnderDragon $this = (EnderDragon) (Object) this;
 		if (source instanceof EntityDamageSource && !((EntityDamageSource)source).isThorns()) {
 			$this.hurt($this.getSubEntities()[2], source, amount);
+		}
+	}
+
+	@Inject(at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/boss/enderdragon/EnderDragon;reallyHurt(Lnet/minecraft/world/damagesource/DamageSource;F)Z", shift = At.Shift.AFTER), method = "hurt(Lnet/minecraft/world/entity/boss/EnderDragonPart;Lnet/minecraft/world/damagesource/DamageSource;F)Z")
+	private void onReallyHurt(EnderDragonPart part, DamageSource damageSource, float amount, CallbackInfoReturnable<Boolean> callbackInfo) {
+		EnderDragon $this = (EnderDragon) (Object) this;
+		if (this.isDeadOrDying() && $this.getPhaseManager().getCurrentPhase().getPhase().equals(CrystalRespawnPhase.getPhaseType())) {
+			$this.setHealth(1.0F);
+			$this.getPhaseManager().setPhase(EnderDragonPhase.DYING);
 		}
 	}
 }